Additional improvements on the jobs and projects page.
Added a header to each actual job and project to separate the content from the header.
Added 'Apply for this job' and 'Get involved' buttons to external locations.
Changed the dates of the posts for the current release.
Made job/project listing items more similar and full block clickable.
Some minor styling changes.