Auto compile Sass files
Previously we had to commit the `assets/stylesheets/site.css` file for
production to use it. It wouldn't compile the file on deploy.

I've upgraded the website to the github-pages gem and updated the config
to auto compile the Sass on deploy. Had to take out compass to do so,
but I think we can do without now. Most things compass is being used for
is now standard in CSS.

And the things that aren't (images used for buttons and ribbons) and be
rewritten in pure CSS.