Move livestream to top of page
When we're live, and have a YouTube video link, show it at the top of
the page. Then when people visit the website they can start watching
immediately.

I wrapped it in an if-statement to toggle if it's visible or not. It's
easier hopefully than wrapping it in HTML comments every time.